+++ src/xmipp/libraries/parallel/mpi_classify_CL2D.cpp
@@ -209,7 +209,7 @@ void CL2DClass::transferUpdate(bool centerReference)
 }
 #undef DEBUG
 
-constexpr float SHIFT_THRESHOLD = 	0.95;		// Shift threshold in pixels.
+constexpr double SHIFT_THRESHOLD = 	0.95;		// Shift threshold in pixels.
 constexpr float ROTATE_THRESHOLD = 	1.0	;		// Rotate threshold in degrees.
 
 constexpr float INITIAL_SHIFT_THRESHOLD = 	SHIFT_THRESHOLD + 1.0;		// Shift threshold in pixels.
